Apply code standards
